Added LoginController in order to separate from the rest of actions.
